chore: broke down controllers into services and functions
